Straight::Order made into a module, removed constructor
This is to make it easier to make it work with ActiveRecord and
any other class, really. You include the module into your class and it works.
If it was a class, we'd have to deal with it being extended by some additional
shitload of code that would handle saving attributes and all that.

Instead, we'll use any ORM we want and any models we want and simply prepend
them with the Straight::Order module.