Use two submodules for (Gateway|Order)Module, Prependable and Includable
The reason for that decision is to not mess with the ancestor methods chain
for those methods that shall not be augmented by the library the module is being included into.
That is, if we include OrderModule into a class that supports saving attributes into
the DB, it is only necessary to prepend getters and setters of the attributes, but
not other methods that are of no concern to that class.

The general rule for the Straight library being that for now we only prepend
getters and setters while everything else is actually included.